+++ package.json
@@ -11,11 +11,18 @@
     "build:all": "run-s build:forc build:connector",
     "build:connector": "pnpm run build --filter=wallet-connector-svm",
     "build": "turbo run build",
+    "lint": "run-s lint:check prettier:check",
+    "lint:check": "eslint . --ext .ts,.tsx,.js,.jsx",
+    "lint:fix": "pnpm lint:check --fix",
     "prettier:check": "prettier --check .",
     "prettier:format": "prettier --write ."
   },
   "devDependencies": {
+    "@fuels/eslint-plugin": "^0.17.0",
     "@fuels/prettier-config": "^0.15.2",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.6",
     "npm-run-all": "^4.1.5",
     "turbo": "^1.11.2"
   }
