+++ packages/wallet-connector-solana/src/index.ts
@@ -1,3 +1,4 @@
-//export * from './SolanaWalletConnector';
-//export * from "./OkxSolanaWalletConnector";
+export * from './SolanaWalletConnector';
 export * from './SolflareSolanaWalletConnector';
+export * from "./utils/predicateUtils";
+
