+++ packages/wallet-connector-solana/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./predicateUtils";
+export * from "./publicKeyUtils";
