+++ packages/wallet-connector-solana/src/utils/publicKeyUtils.ts
@@ -1,0 +1,6 @@
+import type { PublicKey } from "@solana/web3.js";
+import { hexlify } from "fuels";
+
+export const solanaPublicKeyToHex = (publicKey: PublicKey) => {
+  return hexlify(publicKey.toBytes());
+};
